requirements.txt
# Common
django==5.0
python-dotenv==1.0.0

#Authentication 
django-rest-auth==0.9.5

# Rest
djangorestframework==3.14.0

# History Models 
django-simple-history==3.4.0

# Testing
pytest==7.4.3

#Documentation
django-rest-swagger==2.2.0